+++ backend/src/main/java/com/ll/readycode/api/reviews/dto/request/ReviewUpdateRequest.java
@@ -2,6 +2,7 @@ package com.ll.readycode.api.reviews.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public record ReviewUpdateRequest(
     @Schema(description = "수정할 리뷰 내용", example = "내용을 조금 수정했습니다.", maxLength = 1000)
@@ -12,4 +13,4 @@ public record ReviewUpdateRequest(
         @NotNull(message = "평점은 필수입니다.")
         @Min(value = 1, message = "평점은 최소 1점이어야 합니다.")
         @Max(value = 5, message = "평점은 최대 5점이어야 합니다.")
-        Double rating) {}
+        BigDecimal rating) {}
